Usar uint8_t para index, que é 0 ou 1

+++ src/beer_counter.c
@@ -167,7 +167,7 @@ void main(void)
  * é chamada, os displays ligados são alternados. */
 ISR(TIMER1_COMPA_vect)
 {
-    static int idx = 0;
+    static uint8_t idx = 0;
     uint8_t valor_saida = cervejas[!idx] | (cervejas[!idx+2] << 4);
 
     desligar_bit(PORTD, (6 + idx));
